Evaluator rework

+++ gs_init_compare/depth_prediction/utils/download_with_tqdm.py
@@ -1,5 +1,7 @@
+from pathlib import Path
 import urllib.request
 from tqdm import tqdm
+from filelock import FileLock
 
 
 class DownloadProgressBar(tqdm):
@@ -10,14 +12,19 @@ class DownloadProgressBar(tqdm):
 
 
 def download_with_pbar(url, output_path):
-    try:
-        with DownloadProgressBar(
-            unit="B", unit_scale=True, miniters=1, desc=url.split("/")[-1]
-        ) as t:
-            urllib.request.urlretrieve(
-                url, filename=output_path, reporthook=t.update_to
-            )
-    except KeyboardInterrupt:
-        print(f"Keyboard interrupt. Deleting incomplete download at {output_path}")
-        output_path.unlink()
-        raise
+    output_path = Path(output_path)
+    lock_path = output_path.with_suffix(output_path.suffix + ".lock")
+    with FileLock(lock_path):
+        if output_path.exists():
+            return
+        try:
+            with DownloadProgressBar(
+                unit="B", unit_scale=True, miniters=1, desc=url.split("/")[-1]
+            ) as t:
+                urllib.request.urlretrieve(
+                    url, filename=output_path, reporthook=t.update_to
+                )
+        except KeyboardInterrupt:
+            print(f"Keyboard interrupt. Deleting incomplete download at {output_path}")
+            output_path.unlink()
+            raise
